WIP! Moving submission group modal to separate .js file
