+++ bagitobjecttransfer/recordtransfer/views.py
@@ -667,11 +667,17 @@ class TransferFormWizard(SessionWizardView):
                 {
                     "IS_NEW": True,
                     "new_group_form": SubmissionGroupForm(),
-                    "ID_SUBMISSION_GROUP_NAME": ID_SUBMISSION_GROUP_NAME,
-                    "ID_SUBMISSION_GROUP_DESCRIPTION": ID_SUBMISSION_GROUP_DESCRIPTION,
-                    "ID_DISPLAY_GROUP_DESCRIPTION": ID_DISPLAY_GROUP_DESCRIPTION,
-                    "ID_SUBMISSION_GROUP_SELECTION": ID_SUBMISSION_GROUP_SELECTION,
-                    "DEFAULT_GROUP_ID": self.submission_group_uuid,
+                    "js_context": {
+                        "id_submission_group_name": ID_SUBMISSION_GROUP_NAME,
+                        "id_submission_group_description": ID_SUBMISSION_GROUP_DESCRIPTION,
+                        "id_display_group_description": ID_DISPLAY_GROUP_DESCRIPTION,
+                        "id_submission_group_selection": ID_SUBMISSION_GROUP_SELECTION,
+                        "fetch_group_descriptions_url": reverse(
+                            "recordtransfer:get_user_submission_groups",
+                            kwargs={"user_id": self.request.user.pk},
+                        ),
+                        "default_group_id": self.submission_group_uuid,
+                    },
                 }
             )
